bot/requirements.txt
PyGithub
openai==1.102.0
python-dotenv
